Add discounted product price and related changes
This commit adds a new TextField for discounted product price in the `AddProductsScreen`. It also includes changes in the `ProductModel` to include discounted price and discount percentage. The discount amount and discount percentage are calculated and added to the product details before uploading to Firebase. Additionally, the inventory screen now displays the discount percentage for each product.

A new package, `flutter_rating_bar`, has been added to the `pubspec.yaml` file. However, it is not used in the current diff.

The `CustomNavBar` widget has been updated to navigate to the `ProductDisplayScreen` when the search icon is tapped.