Delete categories images and related code
This commit removes the images for the categories Business, Deals, Furniture, More, Movies, Pharmacy, and Travel from the assets/images/categories directory. The related code in constants.dart, users\_product\_services.dart, and main.dart is also removed. The CartScreen widget in CartScreen.dart is also updated to remove the deleted categories. Additionally, the HomeScreen widget in Home\_Screen.dart is updated to remove the CategoriesListView widget and replace it with a horizontal list of categories that navigates to the ProductCategoriesScreen when tapped. The CustomCaroselDeals widget is also added to display the deals of the day.