Update dependencies, change FlutterActivity to FlutterFragmentActivity in MainActivity.kt, update styles.xml and add stripe keys in main.dart
The code diff includes updates to the dependencies in pubspec.yaml, changing FlutterActivity to FlutterFragmentActivity in MainActivity.kt, updating styles.xml, and adding stripe keys in main.dart. The specific changes are as follows:

* Updating flutter\_stripe to version 5.0.0 and adding http as a dependency in pubspec.yaml
* Changing FlutterActivity to FlutterFragmentActivity in MainActivity.kt
* Updating styles.xml to use Theme.MaterialComponents as the parent style
* Adding stripe keys in main.dart

Note: The code diff also includes changes to some other files, but those changes are not relevant to the main updates mentioned above.